Initial commit for image build
PROBLEM:

In order to test out a Vault authentication workflow that includes
Puppet certificates, we need the ability to quickly spin up
a Puppetserver. Having a Docker image for a Puppetserver serves that
need. We're also not interested in using the Puppet-in-Docker
"Pupperware" solution because it's overly complex for what we need.

SOLUTION:

Create a `Dockerfile` for a Debian Docker image that installs
the Puppet/Puppetserver packages and maps the `cakey` and `cacert`
options to the CA cert/key that we've specified.

OUTCOME:

We're able to quickly spin up a Puppetserver instance with our provided
CA cert/key.