Initial build with puppetserver starting
Puppetserver starts, the CA is generated but we don't yet know if
copying the tls.key and tls.crt file into
/etc/puppetlabs/puppetserver/ca-tls works.

We expect Puppetserver to generate the rest of the CA stuff from these
two files, but it might do the wrong thing.

Next step: Publish to quay and test in k8s with a secret resource
mounting the tls.crt and key in the path.